Added CPI-Exception automation
Added FullName property to Person class, which simply returns the full
name of the person (F+M+L, or F+L if no MI)

Added CPIUsers method to the LotusNotesConnection class, which takes a
List of Person object to add to the CPI list. This is the same method as
the "Add" method in CPI_Cleanup_GUI

Added code to PendingUsers class to update the progress bar message, and
invoke the CPIUsers method of it's LotusNotesConnection object.